requirements.txt
cffi==1.17.1
pycparser==2.22
python-qpid-proton==0.40.0
